create UnlockFlag, based on prior extraction of locked flag PR
